[SPOT-286][REFACTOR] Vote 서비스 컨트롤러 분리
