+++ src/main/java/com/example/spot/study/application/StudyMemberCommandServiceImpl.java
@@ -7,12 +7,8 @@ import com.example.spot.common.api.exception.handler.StudyHandler;
 import com.example.spot.member.domain.Member;
 import com.example.spot.report.domain.MemberReport;
 import com.example.spot.notification.domain.Notification;
-import com.example.spot.schedule.domain.repository.QuizRepository;
-import com.example.spot.schedule.domain.repository.QuizSubmissionRepository;
-import com.example.spot.schedule.domain.ScheduleRepository;
 import com.example.spot.story.domain.Story;
 import com.example.spot.story.domain.association.StoryReport;
-import com.example.spot.todo.domain.ToDo;
 import com.example.spot.study.domain.enums.StudyApplicationStatus;
 import com.example.spot.notification.domain.enums.NotifyType;
 import com.example.spot.member.domain.enums.Status;
@@ -25,37 +21,20 @@ import com.example.spot.notification.domain.NotificationRepository;
 import com.example.spot.story.domain.repository.StoryReportRepository;
 import com.example.spot.story.domain.StoryRepository;
 import com.example.spot.study.domain.StudyRepository;
-import com.example.spot.todo.domain.ToDoRepository;
 import com.example.spot.study.presentation.dto.request.StudyHostWithdrawRequestDTO;
 import com.example.spot.study.presentation.dto.request.StudyMemberReportDTO;
-import com.example.spot.study.presentation.dto.request.StudyVoteRequestDTO;
 import com.example.spot.story.web.dto.response.StoryResDTO;
 import com.example.spot.study.presentation.dto.response.StudyTerminationResponseDTO;
-import com.example.spot.study.presentation.dto.response.StudyVoteResponseDTO;
 import com.example.spot.study.presentation.dto.response.StudyWithdrawalResponseDTO;
 import com.example.spot.common.security.utils.SecurityUtils;
 import com.example.spot.common.application.s3.S3ImageService;
 import com.example.spot.member.presentation.dto.MemberResponseDTO;
-import com.example.spot.todo.presentation.dto.request.ToDoListRequestDTO.ToDoListCreateDTO;
-import com.example.spot.todo.presentation.dto.response.ToDoListResponseDTO.ToDoListCreateResponseDTO;
-import com.example.spot.todo.presentation.dto.response.ToDoListResponseDTO.ToDoListUpdateResponseDTO;
 import com.example.spot.study.presentation.dto.response.StudyWithdrawalResponseDTO.WithdrawalDTO;
 import com.example.spot.study.presentation.dto.response.StudyApplyResponseDTO;
 
-import java.util.Objects;
-
-import com.example.spot.vote.domain.Vote;
-import com.example.spot.vote.domain.association.VoteOption;
-import com.example.spot.vote.domain.association.VoteParticipant;
-import com.example.spot.vote.domain.repository.VoteOptionRepository;
-import com.example.spot.vote.domain.repository.VoteParticipantRepository;
-import com.example.spot.vote.domain.VoteRepository;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 
 @Service
@@ -63,24 +42,14 @@ import java.util.List;
 @Transactional
 public class StudyMemberCommandServiceImpl implements StudyMemberCommandService {
 
-    @Value("${image.post.anonymous.profile}")
-    private String defaultImage;
-
     private final MemberRepository memberRepository;
+    private final StudyMemberRepository studyMemberRepository;
     private final StudyRepository studyRepository;
-    private final ScheduleRepository scheduleRepository;
-    private final QuizRepository quizRepository;
-    private final VoteRepository voteRepository;
-    private final VoteOptionRepository voteOptionRepository;
+    private final StoryRepository storyRepository;
+    private final NotificationRepository notificationRepository;
     private final MemberReportRepository memberReportRepository;
     private final StoryReportRepository storyReportRepository;
 
-    private final StudyMemberRepository studyMemberRepository;
-    private final QuizSubmissionRepository quizSubmissionRepository;
-    private final StoryRepository storyRepository;
-    private final VoteParticipantRepository voteParticipantRepository;
-    private final ToDoRepository toDoRepository;
-    private final NotificationRepository notificationRepository;
 
     // S3 Service
     private final S3ImageService s3ImageService;
@@ -301,243 +270,6 @@ public class StudyMemberCommandServiceImpl implements StudyMemberCommandService 
             .build();
     }
 
-/* ----------------------------- 스터디 투표 관련 API ------------------------------------- */
-
-    /**
-     * 스터디 투표를 생성하는 메서드입니다.
-     * @param studyId 타겟 스터디의 아이디를 입력 받습니다.
-     * @param voteDTO 생성할 투표의 제목, 항목 목록, 중복 선택 가능 여부, 종료 일시를 입력 받습니다.
-     * @return 생성된 투표의 아이디와 제목을 반환합니다.
-     */
-    @Override
-    public StudyVoteResponseDTO.VotePreviewDTO createVote(Long studyId, StudyVoteRequestDTO.VoteDTO voteDTO) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        Member loginMember = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        studyMemberRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, StudyApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        //=== Feature ===//
-        Vote vote = Vote.builder()
-                .study(study)
-                .member(loginMember)
-                .title(voteDTO.getTitle())
-                .isMultipleChoice(voteDTO.getIsMultipleChoice())
-                .finishedAt(voteDTO.getFinishedAt())
-                .build();
-
-        // Vote 저장
-        vote = voteRepository.save(vote);
-        // Option 저장
-        vote = createOption(vote, voteDTO);
-        // 연관관계 매핑
-        loginMember.addVote(vote);
-        study.addVote(vote);
-
-        return StudyVoteResponseDTO.VotePreviewDTO.toDTO(vote);
-    }
-
-    /**
-     * 스터디 투표의 항목을 생성하는 메서드입니다.
-     * createVote 메서드 내부에서 사용되는 메서드입니다.
-     * @param vote 항목을 생성할 타겟 투표를 입력 받습니다.
-     * @param voteDTO 생성할 투표의 제목, 항목 목록, 중복 선택 가능 여부, 종료 일시를 입력 받습니다.
-     * @return 투표 객체를 반환합니다.
-     */
-    private Vote createOption(Vote vote, StudyVoteRequestDTO.VoteDTO voteDTO) {
-        voteDTO.getOptions()
-                .forEach(stringOption -> {
-                    VoteOption voteOption = VoteOption.builder()
-                            .vote(vote)
-                            .content(stringOption)
-                            .build();
-                    voteOption = voteOptionRepository.save(voteOption);
-                    vote.addOption(voteOption);
-                });
-        return voteRepository.save(vote);
-    }
-
-    /**
-     * 특정 항목에 투표하기 위한 메서드입니다.
-     * @param studyId 타겟 스터디의 아이디를 입력 받습니다.
-     * @param voteId 타겟 투표의 아이디를 입력 받습니다.
-     * @param votedOptionDTO 회원이 투표한 항목의 아이디 목록을 입력 받습니다.
-     * @return 투표 아이디, 회원 아이디, 투표한 항목 목록을 반환합니다.
-     */
-    @Override
-    public StudyVoteResponseDTO.VotedOptionDTO vote(Long studyId, Long voteId, StudyVoteRequestDTO.VotedOptionDTO votedOptionDTO) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        Member loginMember = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        Vote vote = voteRepository.findById(voteId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_NOT_FOUND));
-        voteRepository.findByIdAndStudyId(voteId, studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        studyMemberRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, StudyApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        // 중복 선택이 허용되지 않는 투표는 여러 개의 option을 선택할 수 없음
-        if (!vote.getIsMultipleChoice() && votedOptionDTO.getOptionIdList().size() > 1) {
-            throw new StudyHandler(ErrorStatus._STUDY_VOTE_MULTIPLE_CHOICE_NOT_VALID);
-        }
-
-        // 한 번 참여한 투표는 다시 참여할 수 없음
-        voteOptionRepository.findAllByVoteId(voteId)
-                .forEach(option -> {
-                    if (voteParticipantRepository.existsByMemberIdAndVoteOptionId(loginMember.getId(), option.getId())) {
-                            throw new StudyHandler(ErrorStatus._STUDY_VOTE_RE_PARTICIPATION_INVALID);
-                    }
-                });
-
-        //=== Feature ===//
-        List<VoteParticipant> voteParticipants = votedOptionDTO.getOptionIdList().stream()
-                .map(optionId -> {
-                    VoteOption votedVoteOption = voteOptionRepository.findById(optionId)
-                            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_OPTION_NOT_FOUND));
-                    voteOptionRepository.findByIdAndVoteId(optionId, voteId)
-                            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_OPTION_NOT_FOUND));
-
-                    VoteParticipant voteParticipant = VoteParticipant.builder()
-                            .member(loginMember)
-                            .voteOption(votedVoteOption)
-                            .build();
-
-                    voteParticipant = voteParticipantRepository.save(voteParticipant);
-                    loginMember.addMemberVote(voteParticipant);
-                    votedVoteOption.addMemberVote(voteParticipant);
-
-                    return voteParticipant;
-                })
-                .toList();
-
-        return StudyVoteResponseDTO.VotedOptionDTO.toDTO(vote, loginMember, voteParticipants);
-    }
-
-    /**
-     * 투표를 편집하는 메서드입니다.
-     * @param studyId 타겟 스터디의 아이디를 입력 받습니다.
-     * @param voteId 편집할 투표의 아이디를 입력 받습니다.
-     * @param voteDTO 편집된 투표의 제목, 항목 목록, 복수 선택 가능 여부, 종료 일시를 입력 받습니다.
-     * @return 편집된 투표의 아이디와 제목을 반환합니다.
-     */
-    @Override
-    public StudyVoteResponseDTO.VotePreviewDTO updateVote(Long studyId, Long voteId, StudyVoteRequestDTO.VoteUpdateDTO voteDTO) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        Member loginMember = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        Vote vote = voteRepository.findById(voteId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        studyMemberRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, StudyApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        // 로그인한 회원이 투표 생성자인지 확인
-        if (!loginMember.equals(vote.getMember())) {
-            throw new StudyHandler(ErrorStatus._STUDY_VOTE_CREATOR_NOT_AUTHORIZED);
-        }
-
-        // 한 명이라도 투표에 참여했으면 투표 편집 불가
-        voteOptionRepository.findAllByVoteId(voteId)
-                .forEach(option -> {
-                    if (voteParticipantRepository.existsByVoteOptionId(option.getId())) {
-                        throw new StudyHandler(ErrorStatus._STUDY_VOTE_IS_IN_PROGRESS);
-                    }
-                });
-
-        //=== Feature ===//
-        for (StudyVoteRequestDTO.OptionDTO optionDTO : voteDTO.getOptions()) {
-            VoteOption voteOption = voteOptionRepository.findById(optionDTO.getOptionId())
-                    .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_OPTION_NOT_FOUND));
-            voteOption.setContent(optionDTO.getContent());
-            voteOption = voteOptionRepository.save(voteOption);
-            vote.updateOption(voteOption);
-        }
-
-        vote.updateVote(voteDTO.getTitle(), voteDTO.getIsMultipleChoice(), voteDTO.getFinishedAt());
-        vote = voteRepository.save(vote);
-        loginMember.updateVote(vote);
-        study.updateVote(vote);
-
-        return StudyVoteResponseDTO.VotePreviewDTO.toDTO(vote);
-    }
-
-    /**
-     * 투표를 삭제하는 메서드입니다.
-     * @param studyId 타겟 스터디의 아이디를 입력 받습니다.
-     * @param voteId 삭제할 투표의 아이디를 입력 받습니다.
-     * @return 삭제된 투표의 아이디와 제목을 반환합니다.
-     */
-    @Override
-    public StudyVoteResponseDTO.VotePreviewDTO deleteVote(Long studyId, Long voteId) {
-
-        //=== Exception ===//
-        Long memberId = SecurityUtils.getCurrentUserId();
-        SecurityUtils.verifyUserId(memberId);
-
-        Member loginMember = memberRepository.findById(memberId)
-                .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-        Study study = studyRepository.findById(studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-        Vote vote = voteRepository.findById(voteId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_NOT_FOUND));
-        voteRepository.findByIdAndStudyId(voteId, studyId)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_VOTE_NOT_FOUND));
-
-        // 로그인한 회원이 스터디 회원인지 확인
-        studyMemberRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, StudyApplicationStatus.APPROVED)
-                .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND));
-
-        // 로그인한 회원이 투표 생성자인지 확인
-        if (!loginMember.equals(vote.getMember())) {
-            throw new StudyHandler(ErrorStatus._STUDY_VOTE_CREATOR_NOT_AUTHORIZED);
-        }
-
-        //=== Feature ===//
-        deleteOptions(voteId);
-        loginMember.deleteVote(vote);
-        study.deleteVote(vote);
-        voteRepository.delete(vote);
-
-        return StudyVoteResponseDTO.VotePreviewDTO.toDTO(vote);
-    }
-
-    /**
-     * 모든 투표 항목을 삭제하는 메서드입니다.
-     * deleteVote 메서드 내부에서 호출되는 메서드입니다.
-     * @param voteId 항목을 삭제할 타겟 투표의 아이디를 입력 받습니다.
-     */
-    private void deleteOptions(Long voteId) {
-        List<VoteOption> voteOptions = voteOptionRepository.findAllByVoteId(voteId);
-        voteOptions.forEach(option -> {
-            option.deleteAllMemberVotes();
-            voteParticipantRepository.deleteAll(voteParticipantRepository.findAllByVoteOptionId(option.getId()));
-            voteOptionRepository.delete(option);
-        });
-    }
-
     /**
      * 회원이 스터디 장인지 확인합니다.
      * @param memberId 확인 하려는 회원 ID
@@ -546,16 +278,6 @@ public class StudyMemberCommandServiceImpl implements StudyMemberCommandService 
      */
     private boolean isOwner(Long memberId, Long studyId) {
         return studyMemberRepository.findByMemberIdAndStudyIdAndIsOwned(memberId, studyId, true).isPresent();
-    }
-
-    /**
-     * 회원이 스터디 구성원인지 확인합니다.
-     * @param memberId 확인 하려는 회원 ID
-     * @param studyId 확인 하려는 스터디 ID
-     * @return 스터디 참여 여부를 반환합니다.
-     */
-    private boolean isMember(Long memberId, Long studyId) {
-        return studyMemberRepository.findByMemberIdAndStudyIdAndStatus(memberId, studyId, StudyApplicationStatus.APPROVED).isPresent();
     }
 
     /**
@@ -643,210 +365,4 @@ public class StudyMemberCommandServiceImpl implements StudyMemberCommandService 
 
         return StoryResDTO.PostPreviewDTO.toDTO(story);
     }
-
-/* ----------------------------- 스터디 To-Do List 관련 API ------------------------------------- */
-
-    /**
-     * To-Do List를 생성합니다.
-     * @param studyId 생성할 To-Do List가 속한 스터디 ID
-     * @param toDoListCreateDTO 생성할 To-Do List 정보
-     * @return 생성된 To-Do List 정보
-     * @throws StudyHandler 스터디를 찾을 수 없을 때
-     * @throws StudyHandler To-Do List를 생성하는 회원이 스터디 회원이 아닐 때
-     * @throws StudyHandler 해당 회원을 찾을 수 없을 때
-     */
-    @Override
-    public ToDoListCreateResponseDTO createToDoList(Long studyId,
-        ToDoListCreateDTO toDoListCreateDTO) {
-
-        // 스터디 조회
-        Study study = studyRepository.findById(studyId)
-            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_NOT_FOUND));
-
-        // To-Do List를 생성하는 회원 ID 조회
-        Long currentUserId = SecurityUtils.getCurrentUserId();
-
-        // To-Do List를 생성하는 회원이 스터디 회원인지 확인
-        if (!isMember(currentUserId, studyId))
-            throw new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND);
-
-        // 회원 조회
-        Member member = memberRepository.findById(currentUserId)
-            .orElseThrow(() -> new MemberHandler(ErrorStatus._MEMBER_NOT_FOUND));
-
-        // To-Do List 생성
-        ToDo toDo = ToDo.builder()
-            .study(study)
-            .member(member)
-            .date(toDoListCreateDTO.getDate())
-            .isDone(false)
-            .content(toDoListCreateDTO.getContent())
-            .build();
-
-        // To-Do List 저장
-        toDo.setToDoList();
-        toDoRepository.save(toDo);
-
-        // To-Do List 생성 DTO 반환
-        return ToDoListCreateResponseDTO.builder()
-            .id(toDo.getId())
-            .content(toDo.getContent())
-            .createdAt(toDo.getCreatedAt())
-            .build();
-    }
-
-    // studyId가 필요할까?
-
-    /**
-     * To-Do List에 작성한 할 일의 체크 상태를 변경 합니다. 체크 상태를 변경 하면 해당 스터디에 참여하고 있는 모든 회원에게 알림이 전송됩니다.
-     * @param studyId 스터디 ID
-     * @param toDoListId 변경할 To-Do List ID
-     * @return To-Do List 변경 여부와 변경 시간
-     * @throws StudyHandler To-Do List를 찾을 수 없을 때
-     * @throws StudyHandler To-Do List가 스터디에 속하지 않을 때
-     * @throws StudyHandler To-Do List를 변경하는 회원이 스터디 회원이 아닐 때
-     * @throws StudyHandler 알림 생성 할 스터디 회원을 찾을 수 없을 때
-     */
-    @Override
-    public ToDoListUpdateResponseDTO checkToDoList(Long studyId, Long toDoListId) {
-
-        // To-Do List 조회
-        ToDo toDo = toDoRepository.findById(toDoListId)
-            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_TODO_NOT_FOUND));
-
-        // To-Do List가 속한 스터디가 아니면 예외 처리
-        if (!Objects.equals(toDo.getStudy().getId(), studyId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_IS_NOT_BELONG_TO_STUDY);
-
-        // To-Do List를 변경하는 회원이 스터디 회원이 아니면 예외 처리
-        Long currentUserId = SecurityUtils.getCurrentUserId();
-        if (!toDo.getMember().getId().equals(currentUserId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_NOT_AUTHORIZED);
-
-        // To-Do List 체크 상태 변경
-        toDo.check();
-
-        // 스터디 회원의 To-Do List 중 하나가 완료 되면, 해당 스터디의 모든 회원에게 알림 전송
-        if (toDo.isDone()){
-            List<Member> members = studyMemberRepository.findAllByStudyIdAndStatus(studyId, StudyApplicationStatus.APPROVED).stream()
-                .map(StudyMember::getMember)
-                .toList();
-
-            // 알림을 생성할 회원이 없으면 알림 생성하지 않음
-            if (members.isEmpty()){
-                return ToDoListUpdateResponseDTO.builder()
-                    .id(toDo.getId())
-                    .isDone(toDo.isDone())
-                    .updatedAt(toDo.getUpdatedAt())
-                    .build();
-            }
-
-            // 알림 생성
-            members.forEach(studyMember -> {
-                Notification notification = Notification.builder()
-                    .member(studyMember)
-                    .notifierName(toDo.getMember().getName()) // To-Do 완료한 회원 이름
-                    .study(toDo.getStudy())
-                    .type(NotifyType.TO_DO_UPDATE)
-                    .isChecked(Boolean.FALSE)
-                    .build();
-                notificationRepository.save(notification);
-            });
-        }
-
-        // To-Do List 저장
-        toDoRepository.save(toDo);
-
-        // To-Do List 변경 DTO 반환
-        return ToDoListUpdateResponseDTO.builder()
-            .id(toDo.getId())
-            .isDone(toDo.isDone())
-            .updatedAt(toDo.getUpdatedAt())
-            .build();
-    }
-
-
-    /**
-     * To-Do List 내용을 수정합니다.
-     * @param studyId 수정할 To-Do List가 속한 스터디 ID
-     * @param toDoListId 수정할 To-Do List ID
-     * @param toDoListCreateDTO 수정할 To-Do List 정보
-     * @return 수정된 To-Do List 정보
-     * @throws StudyHandler To-Do List를 찾을 수 없을 때
-     * @throws StudyHandler To-Do List가 스터디에 속하지 않을 때
-     * @throws StudyHandler To-Do List를 수정하는 회원이 스터디 회원이 아닐 때
-     */
-    @Override
-    public ToDoListUpdateResponseDTO updateToDoList(Long studyId, Long toDoListId,
-        ToDoListCreateDTO toDoListCreateDTO) {
-
-        // To-Do List 조회
-        ToDo toDo = toDoRepository.findById(toDoListId)
-            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_TODO_NOT_FOUND));
-
-        // To-Do List가 속한 스터디가 아니면 예외 처리
-        if (!Objects.equals(toDo.getStudy().getId(), studyId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_IS_NOT_BELONG_TO_STUDY);
-
-        // To-Do List를 수정하는 회원이 스터디 회원이 아니면 예외 처리
-        Long currentUserId = SecurityUtils.getCurrentUserId();
-        if (!toDo.getMember().getId().equals(currentUserId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_NOT_AUTHORIZED);
-
-        // To-Do List 수정
-        toDo.update(toDoListCreateDTO.getContent(), toDoListCreateDTO.getDate());
-
-        // To-Do List 저장
-        toDoRepository.save(toDo);
-
-        // To-Do List 변경 DTO 반환
-        return ToDoListUpdateResponseDTO.builder()
-            .id(toDo.getId())
-            .isDone(toDo.isDone())
-            .updatedAt(toDo.getUpdatedAt())
-            .build();
-    }
-
-    /**
-     * To-Do List를 삭제합니다.
-     * @param studyId 삭제할 To-Do List가 속한 스터디 ID
-     * @param toDoListId 삭제할 To-Do List ID
-     * @return 삭제된 To-Do List 정보
-     * @throws StudyHandler To-Do List를 찾을 수 없을 때
-     * @throws StudyHandler To-Do List가 스터디에 속하지 않을 때
-     * @throws StudyHandler To-Do List를 삭제하는 회원이 스터디 회원이 아닐 때
-     */
-    @Override
-    public ToDoListUpdateResponseDTO deleteToDoList(Long studyId, Long toDoListId) {
-
-        // 로그인 중인 회원 ID 조회
-        Long currentUserId = SecurityUtils.getCurrentUserId();
-
-        // To-Do List를 삭제하는 회원이 스터디 회원인지 확인
-        if (!isMember(currentUserId, studyId))
-            throw new StudyHandler(ErrorStatus._STUDY_MEMBER_NOT_FOUND);
-
-        // To-Do List 조회
-        ToDo toDo = toDoRepository.findById(toDoListId)
-            .orElseThrow(() -> new StudyHandler(ErrorStatus._STUDY_TODO_NOT_FOUND));
-
-        // To-Do List가 속한 스터디가 아니면 예외 처리
-        if (!Objects.equals(toDo.getStudy().getId(), studyId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_IS_NOT_BELONG_TO_STUDY);
-
-        // To-Do List를 삭제하는 회원의 ID와 To-Do List를 생성한 회원의 ID가 다르면 예외 처리
-        if (!toDo.getMember().getId().equals(currentUserId))
-            throw new StudyHandler(ErrorStatus._STUDY_TODO_NOT_AUTHORIZED);
-
-        // To-Do List 삭제
-        toDoRepository.deleteById(toDoListId);
-
-        // To-Do List 삭제 DTO 반환
-        return ToDoListUpdateResponseDTO.builder()
-            .id(toDo.getId())
-            .isDone(toDo.isDone())
-            .updatedAt(toDo.getUpdatedAt())
-            .build();
-    }
-
 }
